Add support to configure Rahmen using a file
We introduce a configuration file that allows to specify the composition
of the status line. The status line consists of fragments derived from
meta data tags, which can be transformed using regular expressions and
capitalization. Currently, a single regular expression can be applied.

In the future, more configuration items are intended to be moved to the
configuration file.

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>